+++ cloudflare.go
@@ -174,25 +174,6 @@ type Owner struct {
 	OwnerType string `json:"owner_type"`
 }
 
-// KeylessSSL represents Keyless SSL configuration.
-type KeylessSSL struct {
-	ID          string    `json:"id"`
-	Name        string    `json:"name"`
-	Host        string    `json:"host"`
-	Port        int       `json:"port"`
-	Status      string    `json:"success"`
-	Enabled     bool      `json:"enabled"`
-	Permissions []string  `json:"permissions"`
-	CreatedOn   time.Time `json:"created_on"`
-	ModifiedOn  time.Time `json:"modifed_on"`
-}
-
-// KeylessSSLResponse represents the response from the Keyless SSL endpoint.
-type KeylessSSLResponse struct {
-	Response
-	Result []KeylessSSL `json:"result"`
-}
-
 // CustomPage represents a custom page configuration.
 type CustomPage struct {
 	CreatedOn      string    `json:"created_on"`
